總評論數、低庫存商品數量get+excel金額取整數

+++ src/main/java/petTopia/dto/shop/OrderItemAnalysisDto.java
@@ -17,7 +17,7 @@ public class OrderItemAnalysisDto {
 	private String productColor;
 	private String productSize;
     private Integer quantity;
-    private BigDecimal unitPrice;
-    private BigDecimal discountPrice;
-    private BigDecimal totalPrice;
+    private double unitPrice;
+    private double discountPrice;
+    private double totalPrice;
 }
